pyproject.toml
[tool.poetry]
package-mode = false
name = "dietician-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.28.1"
python-dotenv = "^1.0.1"
pillow-heif = "^0.16.0"
streamlit = "^1.34.0"
pydantic = "^2.7.1"
typing-extensions = "^4.11.0"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
